requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
django-filter>=24.2
djangorestframework-simplejwt>=5.3,<5.5
drf-spectacular>=0.27,<0.29
psycopg[binary]>=3.1,<3.3
gunicorn>=22,<24
python-dotenv>=1.0,<2
whitenoise>=6.7,<6.10

# Dev / tests
pytest>=8.0,<9
pytest-django>=4.8,<5
pytest-cov>=5.0,<7
factory-boy>=3.3,<4
